initial code from old calendar to new one
This is just the simple outline, there is much to go, but i think it
should be possible to make the calendar
without resourting to images (other than the material
icons that we use for arrows and things.

I am still debating on whether the simple change by a month at a time is sufficient
as most dates are in a narrow band.